+++ src/orquestra/sdk/_base/cli/_repos.py
@@ -598,6 +598,7 @@ def _(wf_run: WorkflowRunSummary) -> ui_models.WFList.WFRow:
         status=wf_run.status.state.value,
         tasks_succeeded=f"{wf_run.completed_task_runs}/{wf_run.total_task_runs}",
         start_time=wf_run.status.start_time,
+        owner=wf_run.owner,
     )